# specs/1-chat-app/contracts/api.yaml
openapi: 3.0.3
info:
  title: Chat App API
  version: 1.0.0
  description: API for web-based chat application
servers:
  - url: http://localhost:3000/api
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
              required:
                - username
                - email
                - password
      responses:
        "201":
          description: User registered
        "400":
          description: Validation error
  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        "401":
          description: Invalid credentials
  /users/contacts:
    get:
      summary: Get user contacts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /groups:
    get:
      summary: Get user groups
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      summary: Create a group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                members:
                  type: array
                  items:
                    type: string
              required:
                - name
                - members
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /messages:
    get:
      summary: Get messages
      security:
        - bearerAuth: []
      parameters:
        - name: recipient
          in: query
          required: true
          schema:
            type: string
        - name: recipientType
          in: query
          required: true
          schema:
            type: string
            enum: [user, group]
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      summary: Send a message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipient:
                  type: string
                recipientType:
                  type: string
                  enum: [user, group]
                content:
                  type: string
                attachments:
                  type: array
                  items:
                    type: string
              required:
                - recipient
                - recipientType
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{id}/reactions:
    post:
      summary: Add reaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                emoji:
                  type: string
              required:
                - emoji
      responses:
        "201":
          description: Reaction added
  /messages/{id}/pin:
    post:
      summary: Pin message
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Message pinned
  /files:
    post:
      summary: Upload file
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        avatar:
          type: string
    Group:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        members:
          type: array
          items:
            type: string
        createdBy:
          type: string
    Message:
      type: object
      properties:
        id:
          type: string
        sender:
          type: string
        recipient:
          type: string
        recipientType:
          type: string
        content:
          type: string
        attachments:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
        reactions:
          type: array
          items:
            type: string
        pinned:
          type: boolean
    File:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        url:
          type: string
        size:
          type: integer
        mimeType:
          type: string